Web UI query calls into real tensor prediction
* Use express-proxy to target the internal tensor server
* move predictionServer into module, make it an object
* use main.py to start training and server


Also rename itemsClient to apiClient
