+++ js/package.json
@@ -27,9 +27,10 @@
     "cheerio": "^0.22.0",
     "classnames": "^2.2.5",
     "express": "^4.13.3",
+    "express-http-proxy": "^0.10.1",
     "html-webpack-plugin": "^2.22.0",
     "moment": "^2.15.0",
-    "mongodb": "2.1.18",
+    "mongodb": "2.2.19",
     "mongoose": "4.5.9",
     "natural": "^0.4.0",
     "nodemon": "^1.10.2",
@@ -47,7 +48,7 @@
     "socket.io-client": "^1.4.8",
     "url": "^0.11.0",
     "url-loader": "^0.5.6",
-    "uuid": "2.0.2",
+    "uuid": "3.0.1",
     "winston": "^2.2.0"
   },
   "devDependencies": {
