+++ js/src/server-fvt/apiFvt.js
@@ -1,5 +1,5 @@
 const expect = require('chai').expect
-const ItemClient = require('../common/itemClient')
+const ItemClient = require('../common/apiClient')
 const uuid = require('uuid')
 var client = new ItemClient()
