+++ tensor/Dockerfile
@@ -3,4 +3,6 @@ FROM tensorflow
 ADD requirements.txt .
 RUN pip3 install -r requirements.txt
 ADD . .
-CMD ["./train.py"]
+
+EXPOSE 8080
+CMD ["./main.py"]
