--- tensor/docker-compose.yml
@@ -1,9 +1,0 @@
-version: '2'
-services:
-  prediction:
-    build: .
-    image: prediction
-    volumes:
-     - training:/work/runs
-volumes:
-  training: {}
